machine-learning-client/requirements123.txt
-i https://pypi.org/simple

certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_version >= '3.7.0'
gitcommitlogger==1.2.5; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.2.1; python_version >= '3.8'

torch==2.1.0; python_version >= '3.6'
torchvision==0.16.0; python_version >= '3.6'
Pillow==9.5.0; python_version >= '3.6'
pytest==7.4.4; python_version >= '3.6'
pymongo>=3.12.0
numpy<2
flask
